install successfully on emulator

+++ app/src/main/java/com/pic/shot/picker/PhotoPagerActivity.java
@@ -23,7 +23,7 @@ public class PhotoPagerActivity extends AppCompatActivity {
 
     public void onCreate(Bundle bundle) {
         super.onCreate(bundle);
-        setContentView((int) R.layout.activity_photo_pager);
+        setContentView(R.layout.activity_photo_pager);
         int intExtra = getIntent().getIntExtra(PhotoPreview.EXTRA_CURRENT_ITEM, 0);
         ArrayList<String> stringArrayListExtra = getIntent().getStringArrayListExtra(PhotoPreview.EXTRA_PHOTOS);
         this.showDelete = getIntent().getBooleanExtra(PhotoPreview.EXTRA_SHOW_DELETE, true);
